+++ test/io_tests/secure_session_test.py
@@ -13,13 +13,14 @@ from xknx.knxip import (
     HPAI,
     KNXIPFrame,
     SecureWrapper,
+    SessionAuthenticate,
     SessionRequest,
     SessionResponse,
     SessionStatus,
 )
 from xknx.knxip.knxip_enum import SecureSessionStatusCode
 
-from ..conftest import EventLoopClockAdvancer
+from ..conftest import EventLoopClockAdvancer, skip_3_10
 
 
 class TestSecureSession:
@@ -71,14 +72,21 @@ class TestSecureSession:
         self.patch_serial_number.stop()
         self.patch_message_tag.stop()
 
+    @skip_3_10
     @patch("xknx.io.transport.tcp_transport.TCPTransport.connect")
     @patch("xknx.io.transport.tcp_transport.TCPTransport.send")
     @patch(
         "xknx.io.ip_secure.generate_ecdh_key_pair",
         return_value=(mock_private_key, mock_public_key),
     )
+    @patch(
+        "xknx.io.ip_secure.SecureSession.send",
+        wraps=SecureSession.send,
+        autospec=True,
+    )
     async def test_lifecycle(
         self,
+        mock_session_send: Mock,
         _mock_generate: Mock,
         mock_super_send: Mock,
         mock_super_connect: Mock,
@@ -92,7 +100,12 @@ class TestSecureSession:
         session_request_frame = KNXIPFrame.init_from_body(
             SessionRequest(ecdh_client_public_key=self.mock_public_key)
         )
-        mock_super_send.assert_called_once_with(
+        mock_session_send.assert_called_once_with(  # unencrypted
+            self.session,  # account for self argument in wraps
+            session_request_frame,
+        )
+        mock_session_send.reset_mock()
+        mock_super_send.assert_called_once_with(  # unencrypted
             session_request_frame,
             None,  # None for addr in TCP transport
         )
@@ -110,6 +123,19 @@ class TestSecureSession:
         self.session.handle_knxipframe(session_response_frame, HPAI(*self.mock_addr))
         await time_travel(0)
         # outgoing
+        authenticate_frame = KNXIPFrame.init_from_body(
+            SessionAuthenticate(
+                user_id=self.mock_user_id,
+                message_authentication_code=bytes.fromhex(
+                    "1f 1d 59 ea 9f 12 a1 52 e5 d9 72 7f 08 46 2c de"
+                ),
+            )
+        )
+        mock_session_send.assert_called_once_with(
+            self.session,  # account for self argument in wraps
+            authenticate_frame,
+        )
+        mock_session_send.reset_mock()
         encrypted_authenticate_frame = KNXIPFrame.init_from_body(
             SecureWrapper(
                 secure_session_id=self.mock_session_id,
@@ -164,16 +190,20 @@ class TestSecureSession:
 
         # keepalive SessionStatus (not specific for sake of simplicity)
         await time_travel(SESSION_KEEPALIVE_RATE)
-        mock_super_send.assert_called_once()
+        mock_session_send.assert_called_once()  # unencrypted
+        mock_session_send.reset_mock()
+        mock_super_send.assert_called_once()  # encrypted
         mock_super_send.reset_mock()
 
         # SessionStatus CLOSE sent on graceful disconnect
         with (
-            patch.object(self.session, "send", wraps=self.session.send) as mock_send,
             patch.object(self.session, "transport") as mock_transport,
         ):
             self.session.stop()
-            mock_send.assert_called_once_with(session_status_close_frame)
+            mock_session_send.assert_called_once_with(
+                self.session,  # account for self argument in wraps
+                session_status_close_frame,
+            )
             mock_super_send.assert_called_once()
             mock_transport.close.assert_called_once()
             assert self.session._keepalive_task is None